Rename requests to avoid collision
