+++ pulsar/client/transport/__init__.py
@@ -2,7 +2,7 @@ import os
 
 from .curl import curl_available
 from .curl import PycurlTransport
-from .requests import requests_multipart_post_available
+from ._requests import requests_multipart_post_available
 from .ssh import rsync_get_file, scp_get_file
 from .ssh import rsync_post_file, scp_post_file
 from .standard import Urllib2Transport
